requirements.txt
langchain>=0.1.0
langgraph>=0.0.10
requests>=2.31.0
pydantic>=2.4.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.0
jinja2>=3.1.0
